Add files via upload
I believe MySQL isn't allowing me to export my tables. I don't know if it's an error on my end or if it's a MySQL error, but I've added screenshots to show you that the database works and that I completed the assignment. I hope the csv files and the SQL file work on your end, but if not please let me know so I can fix it.