Add files via upload
I was able to find out the problem with why I couldn't export, it was because since I have an older Mac, I downloaded an older version of MySQL, unfortunately, there's an error with the export feature in that version (https://bugs.mysql.com/bug.php?id=103783) 

I was able to do it off of VS Code like you suggested and it works now! Thank you and I hope you enjoy my project.